map core dtypes to np dtypes in NumpyFeature

+++ temporian/core/test/utils.py
@@ -27,8 +27,12 @@ class OpO1(base.Operator):
             "output",
             Event(
                 features=[
-                    Feature("f1", dtype.FLOAT, sampling=sampling, creator=self),
-                    Feature("f2", dtype.FLOAT, sampling=sampling, creator=self),
+                    Feature(
+                        "f1", dtype.FLOAT64, sampling=sampling, creator=self
+                    ),
+                    Feature(
+                        "f2", dtype.FLOAT64, sampling=sampling, creator=self
+                    ),
                 ],
                 sampling=sampling,
             ),
@@ -57,7 +61,7 @@ class OpI1O1(base.Operator):
                 features=[
                     Feature(
                         "f1",
-                        dtype.FLOAT,
+                        dtype.FLOAT64,
                         sampling=event.sampling(),
                         creator=self,
                     )
@@ -93,7 +97,7 @@ class OpI2O1(base.Operator):
                 features=[
                     Feature(
                         "f1",
-                        dtype.FLOAT,
+                        dtype.FLOAT64,
                         sampling=event_1.sampling(),
                         creator=self,
                     )
@@ -130,7 +134,7 @@ class OpI1O2(base.Operator):
                 features=[
                     Feature(
                         "f1",
-                        dtype.FLOAT,
+                        dtype.FLOAT64,
                         sampling=event.sampling(),
                         creator=self,
                     )
@@ -144,7 +148,7 @@ class OpI1O2(base.Operator):
                 features=[
                     Feature(
                         "f1",
-                        dtype.FLOAT,
+                        dtype.FLOAT64,
                         sampling=event.sampling(),
                         creator=self,
                     )
